pythonWeb/walletconnect.py
from pkcs7 import PKCS7Encoder
from hashlib import md5
from base64 import b64decode
from base64 import b64encode

from Crypto.Cipher import AES
from Crypto.Random import get_random_bytes
from Crypto.Util.Padding import pad, unpad

import json
import asyncio
import websockets