pythonWeb/requirements.txt
Flask~=1.1
pkcs7
pycryptodome
websockets==8.1